no default exports, move networkAvailability function to generateMetricsReport.ts
